Move metadata to top level.
